Message box when pdfs are saved

+++ utils/clearOrders.js
@@ -8,9 +8,12 @@ module.exports = () => {
 	fs.readdir(directory, (err, files) => {
 		if (err) throw err;
 		for (const file of files) {
-			fs.unlink(path.join(directory, file), err => {
-				if (err) throw err;
-			});
+			const stats = fs.statSync(path.join(directory, file));
+			if (stats.isFile()) {
+				fs.unlink(path.join(directory, file), err => {
+					if (err) throw err;
+				});
+			}
 		}
 	});
 };
